fixed all errors except in modules to be removed
