+++ zingolib/src/data.rs
@@ -1,6 +1,5 @@
 //! This is a mod for data structs that will be used across all sections of zingolib.
 pub mod proposal;
-pub mod witness_trees;
 
 /// transforming data related to the destination of a send.
 pub mod receivers {
